requirements.txt
fastapi[all]==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
sqlalchemy==2.0.36
aiomysql==0.2.0
aiosqlite==0.22.1
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
pydantic-settings==2.5.2
pycryptodome==3.21.0
redis==5.2.0
celery==5.4.0
greenlet==3.3.2
pymysql==1.1.1
